feat: coverage improvments
